requirements.txt
anyio==3.6.2
attrs==22.2.0
autopep8==2.0.1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
coverage==7.0.1
cryptography==38.0.4
dnspython==2.2.1
ecdsa==0.18.0
exceptiongroup==1.1.0
fastapi==0.88.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
packaging==22.0
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pymongo==4.3.3
pytest==7.2.0
pytest-cov==4.0.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
rfc3986==1.5.0
rsa==4.9
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing-extensions==4.4.0
uvicorn==0.20.0
